Feature: Add master_setup command.
For EMR runners, this new option provides functionality to run setup
commands on the master node prior to running a step on EMR. This is
needed in particular for spark jobs. There is a command `setup` that
will perform a similar job, but as the documentation states "you
can’t wrap Spark jobs in setup scripts".

Example of use:

```
runners:
  emr:

    cmdenv:
      JOB_ENV: something

    bootstrap:
      - <some initial cluster configuration>

    master_setup:
      - sudo mkdir -p /tmp/$SOMETHING
      - sudo cp code.tar.gz /tmp/$JOB_ENV
      - <distributes code in cluster>

    upload_files:
      - code.tar.gz#code.tar.gz
```