YARM EMR Runner: Parameterized cmdenv.
Supports a single format string in each `cmdenv` which will be parameterized
based on the job run. This will allow us to run multiple jobs in the cluster in
separate virtual environments. This necessary to handle code updates. This
solution probably is not the best way to handle this, but is an easy initial
solution.