Feature: Add bypass_pool_wait and scheduling queue.
We add a new option `bypass_pool_wait` that if set will ignore skip the pool
wait (set via `pool_wait_minutes`) if there are no clusters with a valid setup.
Previously, if we made a config change that requires new clusters we would have
to wait the full wait time (15-20 minutes) until we provision a new cluster.
This new option bypasses such a wait.

We also need to add a light-weight scheduling queue, lest we face a thundering
herd of clusters all starting at the same time. This uses S3 to provide basic
queueing logic.